requirements.txt
streamlit==1.28.0
spacy==3.7.2
python-docx==0.8.11
PyPDF2==3.0.1
pdfminer.six==20221105
fpdf2==2.7.0
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3
requests==2.31.0
